refactor: split query output surfaces by concern
Move summary/list/stats rendering and stream rendering into dedicated CLI modules so query_output.py remains the coordinator for result delivery, browser, and clipboard paths.
