+++ polylogue/cli/query_summary_output.py
@@ -1,0 +1,344 @@
+"""Summary/list/stats output helpers for query execution."""
+
+from __future__ import annotations
+
+import json
+from typing import TYPE_CHECKING
+
+import click
+
+from polylogue.cli.query_helpers import summary_to_dict
+
+if TYPE_CHECKING:
+    from polylogue.cli.types import AppEnv
+    from polylogue.lib.filters import ConversationFilter
+    from polylogue.lib.models import Conversation, ConversationSummary
+    from polylogue.storage.repository import ConversationRepository
+
+
+def format_summary_list(
+    summaries: list[ConversationSummary],
+    output_format: str,
+    fields: str | None,
+    *,
+    message_counts: dict[str, int] | None = None,
+) -> str:
+    """Format summary-list output for deterministic machine/plain surfaces."""
+    message_counts = message_counts or {}
+
+    selected: set[str] | None = None
+    if fields:
+        selected = {field.strip() for field in fields.split(",")}
+
+    data = [summary_to_dict(summary, message_counts.get(str(summary.id), 0)) for summary in summaries]
+    if selected:
+        data = [{key: value for key, value in item.items() if key in selected} for item in data]
+
+    if output_format == "json":
+        return json.dumps(data, indent=2)
+
+    if output_format == "yaml":
+        import yaml
+
+        return yaml.dump(data, default_flow_style=False, allow_unicode=True)
+
+    if output_format == "csv":
+        import csv
+        import io
+
+        buf = io.StringIO()
+        writer = csv.writer(buf)
+        writer.writerow(["id", "date", "provider", "title", "messages", "tags", "summary"])
+        for summary in summaries:
+            date = summary.display_date.strftime("%Y-%m-%d") if summary.display_date else ""
+            tags_str = ",".join(summary.tags) if summary.tags else ""
+            writer.writerow([
+                str(summary.id),
+                date,
+                summary.provider,
+                summary.display_title or "",
+                message_counts.get(str(summary.id), 0),
+                tags_str,
+                summary.summary or "",
+            ])
+        return buf.getvalue().rstrip("\r\n")
+
+    lines = []
+    for summary in summaries:
+        date = summary.display_date.strftime("%Y-%m-%d") if summary.display_date else ""
+        raw_title = summary.display_title or str(summary.id)[:20]
+        title = (raw_title[:47] + "...") if len(raw_title) > 50 else raw_title
+        count = message_counts.get(str(summary.id), 0)
+        lines.append(f"{str(summary.id)[:24]:24s}  {date:10s}  {summary.provider:12s}  {title} ({count} msgs)")
+    return "\n".join(lines)
+
+
+async def output_stats_sql(
+    env: AppEnv,
+    filter_chain: ConversationFilter,
+    repo: ConversationRepository,
+) -> None:
+    """Output statistics using SQL aggregation without full message loading."""
+    from datetime import datetime, timezone
+
+    has_filters = bool(filter_chain.describe())
+
+    if has_filters:
+        summaries = await filter_chain.list_summaries() if filter_chain.can_use_summaries() else None
+        if summaries is not None:
+            if not summaries:
+                env.ui.console.print("No conversations matched.")
+                return
+            conv_ids = [str(summary.id) for summary in summaries]
+            conv_count = len(conv_ids)
+        else:
+            conv_count = await filter_chain.count()
+            if conv_count == 0:
+                env.ui.console.print("No conversations matched.")
+                return
+            conv_ids = None
+    else:
+        conv_ids = None
+        conv_count = await filter_chain.count()
+        if conv_count == 0:
+            env.ui.console.print("No conversations in archive.")
+            return
+
+    stats = await repo.queries.aggregate_message_stats(conv_ids)
+
+    date_range = ""
+    if stats["min_sort_key"] and stats["max_sort_key"]:
+        min_date = datetime.fromtimestamp(stats["min_sort_key"], tz=timezone.utc).strftime("%Y-%m-%d")
+        max_date = datetime.fromtimestamp(stats["max_sort_key"], tz=timezone.utc).strftime("%Y-%m-%d")
+        date_range = f"{min_date} to {max_date}"
+
+    out = env.ui.console.print
+    out(f"\nConversations: {conv_count:,}\n")
+    if stats["user"] or stats["assistant"]:
+        out(
+            f"Messages: {stats['total']:,} total "
+            f"({stats['user']:,} user, {stats['assistant']:,} assistant)"
+        )
+    else:
+        out(f"Messages: {stats['total']:,}")
+
+    if stats["words_approx"]:
+        out(f"Words: ~{stats['words_approx']:,}")
+
+    if stats.get("providers"):
+        provider_parts = [f"{name} ({count:,})" for name, count in stats["providers"].items()]
+        out(f"Providers: {', '.join(provider_parts)}")
+
+    out(f"Attachments: {stats['attachments']:,}")
+    if date_range:
+        out(f"Date range: {date_range}")
+
+
+def output_stats_by_summaries(
+    env: AppEnv,
+    summaries: list[ConversationSummary],
+    msg_counts: dict[str, int],
+    dimension: str,
+) -> None:
+    """Fast stats-by using lightweight summaries and precomputed message counts."""
+    from collections import defaultdict
+
+    from rich.table import Table
+
+    from polylogue.ui.theme import provider_color
+
+    if not summaries:
+        env.ui.console.print("No conversations matched.")
+        return
+
+    groups: dict[str, list[ConversationSummary]] = defaultdict(list)
+    for summary in summaries:
+        if dimension == "provider":
+            key = str(summary.provider) if summary.provider else "unknown"
+        elif dimension == "month":
+            dt = summary.updated_at or summary.created_at
+            key = dt.strftime("%Y-%m") if dt else "unknown"
+        elif dimension == "year":
+            dt = summary.updated_at or summary.created_at
+            key = dt.strftime("%Y") if dt else "unknown"
+        elif dimension == "day":
+            dt = summary.updated_at or summary.created_at
+            key = dt.strftime("%Y-%m-%d") if dt else "unknown"
+        else:
+            key = "all"
+        groups[key].append(summary)
+
+    sorted_keys = sorted(groups.keys(), reverse=True) if dimension in {"month", "year", "day"} else sorted(groups.keys())
+
+    env.ui.console.print(f"\nMatched: {len(summaries)} conversations (by {dimension})\n")
+
+    table = Table(show_header=True, header_style="bold", box=None, pad_edge=False)
+    table.add_column("Group", style="bold", min_width=12)
+    table.add_column("Convs", justify="right")
+    table.add_column("Messages", justify="right")
+
+    total_convs = 0
+    total_msgs = 0
+
+    for key in sorted_keys:
+        group_summaries = groups[key]
+        n_convs = len(group_summaries)
+        n_msgs = sum(msg_counts.get(str(summary.id), 0) for summary in group_summaries)
+        label = f"[{provider_color(key).hex}]{key}[/]" if dimension == "provider" else key
+        table.add_row(label, f"{n_convs:,}", f"{n_msgs:,}")
+        total_convs += n_convs
+        total_msgs += n_msgs
+
+    table.add_section()
+    table.add_row("[bold]TOTAL[/]", f"[bold]{total_convs:,}[/]", f"[bold]{total_msgs:,}[/]")
+
+    env.ui.console.print(table)
+
+
+def output_stats_by_conversations(
+    env: AppEnv,
+    results: list[Conversation],
+    dimension: str,
+) -> None:
+    """Output grouped statistics from fully hydrated conversations."""
+    from collections import defaultdict
+
+    from rich.table import Table
+
+    from polylogue.ui.theme import provider_color
+
+    if not results:
+        env.ui.console.print("No conversations matched.")
+        return
+
+    groups: dict[str, list[Conversation]] = defaultdict(list)
+    for conv in results:
+        if dimension == "provider":
+            key = conv.provider or "unknown"
+        elif dimension == "month":
+            dt = conv.display_date
+            key = dt.strftime("%Y-%m") if dt else "unknown"
+        elif dimension == "year":
+            dt = conv.display_date
+            key = dt.strftime("%Y") if dt else "unknown"
+        elif dimension == "day":
+            dt = conv.display_date
+            key = dt.strftime("%Y-%m-%d") if dt else "unknown"
+        else:
+            key = "all"
+        groups[key].append(conv)
+
+    sorted_keys = sorted(groups.keys(), reverse=True) if dimension in {"month", "year", "day"} else sorted(groups.keys())
+
+    env.ui.console.print(f"\nMatched: {len(results)} conversations (by {dimension})\n")
+
+    table = Table(show_header=True, header_style="bold", box=None, pad_edge=False)
+    table.add_column("Group", style="bold", min_width=12)
+    table.add_column("Convs", justify="right")
+    table.add_column("Messages", justify="right")
+    table.add_column("Words", justify="right")
+
+    total_convs = 0
+    total_msgs = 0
+    total_words = 0
+
+    for key in sorted_keys:
+        convs = groups[key]
+        n_convs = len(convs)
+        n_msgs = sum(len(conv.messages) for conv in convs)
+        n_words = sum(sum(message.word_count for message in conv.messages) for conv in convs)
+        label = f"[{provider_color(key).hex}]{key}[/]" if dimension == "provider" else key
+        table.add_row(label, f"{n_convs:,}", f"{n_msgs:,}", f"{n_words:,}")
+        total_convs += n_convs
+        total_msgs += n_msgs
+        total_words += n_words
+
+    table.add_section()
+    table.add_row("[bold]TOTAL[/]", f"[bold]{total_convs:,}[/]", f"[bold]{total_msgs:,}[/]", f"[bold]{total_words:,}[/]")
+
+    env.ui.console.print(table)
+
+
+async def output_summary_list(
+    env: AppEnv,
+    summaries: list[ConversationSummary],
+    params: dict[str, object],
+    repo: ConversationRepository | None = None,
+) -> None:
+    """Output a list of conversation summaries with optional rich table rendering."""
+    output_format = str(params.get("output_format", "text"))
+    msg_counts: dict[str, int] = {}
+    if repo:
+        ids = [str(summary.id) for summary in summaries]
+        msg_counts = await repo.queries.get_message_counts_batch(ids)
+
+    fields = params.get("fields")
+    fields_value = str(fields) if isinstance(fields, str) else None
+    if output_format in {"json", "yaml", "csv"} or env.ui.plain:
+        click.echo(
+            format_summary_list(
+                summaries,
+                "text" if env.ui.plain and output_format not in {"json", "yaml", "csv"} else output_format,
+                fields_value,
+                message_counts=msg_counts,
+            )
+        )
+        return
+
+    from rich.table import Table
+    from rich.text import Text
+
+    from polylogue.ui.theme import provider_color
+
+    table = Table(show_header=True, header_style="bold", box=None, pad_edge=False, show_edge=False)
+    table.add_column("ID", style="dim", max_width=24, no_wrap=True)
+    table.add_column("Date", style="dim")
+    table.add_column("Provider")
+    table.add_column("Title", ratio=1)
+    table.add_column("Msgs", justify="right")
+
+    for summary in summaries:
+        date = summary.display_date.strftime("%Y-%m-%d") if summary.display_date else ""
+        raw_title = summary.display_title or str(summary.id)[:20]
+        title = (raw_title[:60] + "...") if len(raw_title) > 63 else raw_title
+        count = msg_counts.get(str(summary.id), 0)
+        provider_text = Text(summary.provider, style=provider_color(summary.provider).hex)
+        table.add_row(str(summary.id)[:24], date, provider_text, title, str(count))
+
+    env.ui.console.print(table)
+
+
+def conversations_to_csv(results: list[Conversation]) -> str:
+    """Convert hydrated conversations to CSV."""
+    import csv
+    import io
+
+    output = io.StringIO()
+    writer = csv.writer(output)
+    writer.writerow(["id", "date", "provider", "title", "messages", "words", "tags", "summary"])
+
+    for conv in results:
+        date = conv.display_date.strftime("%Y-%m-%d") if conv.display_date else ""
+        tags_str = ",".join(conv.tags) if conv.tags else ""
+        writer.writerow([
+            str(conv.id),
+            date,
+            conv.provider,
+            conv.display_title or "",
+            len(conv.messages),
+            sum(message.word_count for message in conv.messages),
+            tags_str,
+            conv.summary or "",
+        ])
+
+    return output.getvalue()
+
+
+__all__ = [
+    "conversations_to_csv",
+    "format_summary_list",
+    "output_stats_by_conversations",
+    "output_stats_by_summaries",
+    "output_stats_sql",
+    "output_summary_list",
+]
